# [BINS-515] xcresult to junit conversion optimization